Implements tasks that stay in order
Since the tasks were not staying in order, I had to wildly change
the way that the program worked. It now puts all the tasks in
from local storage in the right way. It also keeps local storage
clean from trivial bits of data.

This also adds a bunch of missing semicolons in. It also
removes the option to add blank tasks to the task list.

The index.html change is just toadd a comment to let the
you know where tasks will be added since they are added
under the ul tag.