Incorporates changable text
In order to change the task after creating the task, the task
name should be editable. This does that by allowing the user
to select the text and edit it. By making the text editable,
it does remove the ability to delete though, so that is why this
change is in a new branch.